requirements.txt
groq>=0.4.1
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
redis>=5.0.1
chromadb>=0.4.18
sentence-transformers>=2.2.2
httpx>=0.25.2
python-multipart>=0.0.6
python-dotenv>=1.0.0
asyncio-throttle>=1.0.2
tenacity>=8.2.3
PyPDF2>=3.0.1
pdfplumber>=0.10.0
Pillow>=10.0.0
pytesseract>=0.3.10